Move fundNewProperty() into OmniTestSupport
